OpenGLRenderer::added support for 2d sprites for simple 2d visualizations (mountain-car was pleading for one)

+++ RLSimion/worlds/templatedConfigFile.cpp
@@ -31,7 +31,7 @@ bool TemplatedConfigFile::load(const char* inTemplateConfigFile)
 	fopen_s(&templateFile, inTemplateConfigFile, "r");
 	if (templateFile)
 	{
-		numCharsRead = fread_s(m_pTemplateConfigFileContent, MAX_CONFIG_FILE_SIZE, (int)sizeof(char)
+		numCharsRead = fread_s(m_pTemplateConfigFileContent, MAX_CONFIG_FILE_SIZE, sizeof(char)
 			, MAX_CONFIG_FILE_SIZE, templateFile);
 		m_pTemplateConfigFileContent[numCharsRead] = 0;
